// frontend/src/api/generated/model/resourceHealthEventResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CloudOps API
 * Infrastructure resource management, monitoring and asynchronous tasks. Timestamps use UTC ISO-8601. Error responses share ApiError; errors contains field violations.
 * OpenAPI spec version: 0.0.1-SNAPSHOT
 */
import type { ResourceHealthEventResponseFromStatus } from "./resourceHealthEventResponseFromStatus";
import type { ResourceHealthEventResponseToStatus } from "./resourceHealthEventResponseToStatus";

export interface ResourceHealthEventResponse {
  id?: number;
  fromStatus?: ResourceHealthEventResponseFromStatus;
  toStatus?: ResourceHealthEventResponseToStatus;
  changedAt?: string;
}